+++ doc/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Version 3.7 (TBA)
+* Added: New option ` --crawl-exclude` regarding setting regular expression for excluding pages from crawling (e.g. "logout").
 * Revised: Improvement regarding `--crawl` option, for skipping further tests involving target that an injection point has already been detected.
 * Added: Support regarding combining `--crawl` option with scanning multiple targets given from piped-input (i.e. stdin).
 * Revised: Minor improvement regarding adding PCRE_REPLACE_EVAL `/e` modifier (i.e. dynamic code evaluation technique).